Added Usr2/4/6/8 tests and can do BDD test now

+++ src/PlaceItControllers/PlaceItController.java
@@ -39,8 +39,8 @@ public class PlaceItController {
 		PlaceIt placeit = new PlaceIt(titleText, descText, position.latitude,
 				position.longitude);
 		
-		int insertId = (int) db.addPlaceIt(placeit);
-		placeit.setID(insertId);
+		long insertId = db.addPlaceIt(placeit);
+		placeit.setID((int) insertId);
 		placeits.add(placeit);
 		view.addMarker(placeit);
 		return placeit;
